manylinux-build-wheels: Configure project with correct options
* To avoid re-downloading the source, pass SimpleITK_SOURCE_DIR

* Specify correct include_dir/library/interpreter.

* Finally, since the python interpreter exports its symbol (see [1]), SimpleITK python
  modules should not link against any python libraries.
  To ensure it is not the case, we configure the project using an empty
  file as python library.


[1] "Note that libpythonX.Y.so.1 is not on the list of libraries that
a manylinux1 extension is allowed to link to. Explicitly linking to
libpythonX.Y.so.1 is unnecessary in almost all cases: the way ELF linking
works, extension modules that are loaded into the interpreter automatically
get access to all of the interpreter's symbols, regardless of whether or
not the extension itself is explicitly linked against libpython. [...]"

Source: https://www.python.org/dev/peps/pep-0513/#libpythonx-y-so-1